requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
requests==2.31.0
PyMuPDF==1.23.8
python-dotenv==1.0.1
